Use Integer instead of Long for entity IDs (match DB schema)

+++ backend/src/main/java/com/reservex/backend/controllers/AuthController.java
@@ -43,6 +43,6 @@ public class AuthController {
         }
     }
 
-    public record RegisterResponse(Long id, String email, String businessName) {}
+    public record RegisterResponse(Integer id, String email, String businessName) {}
     public record ErrorMessage(String message) {}
 }
