+++ backend/src/main/java/com/reservex/backend/repositories/StallRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface StallRepository extends JpaRepository<Stall, Long> {
+public interface StallRepository extends JpaRepository<Stall, Integer> {
 
     List<Stall> findAllByOrderByNameAsc();
 }
